extension: we need host_permissions to fetch from random websites
Apparently, firefox on linux can fetch regardless of host_permissions.
But firefox on mac and android needs host_permissions otherwise the
fetch fails.
Requesting the permission is seemingly enough to get them (but that
might be an artifact of the testing, since installing an extension
on android via the dev interface might grant more permissions than
a regular install would).

For chrome, host_permissions is not necessary on mac or or linux.
But now that we request it, we generate a user prompt. Oh well, seems
good enough.